ong_trading/ML/requirements.txt
# Tensorflow in mac m1 (follow https://developer.apple.com/metal/tensorflow-plugin/):
# conda install -c apple tensorflow-deps
tensorflow-macos
tensorflow-metal
gym[all]        # open-ai gym
sklearn
seaborn     # nice plots
ta-lib      # If it does not work on M1, then use 'arch -arm64 brew install ta-lib'
